refactor: ScheduleGrid 컴포넌트 분리 및 최적화
