+++ src/components/ScheduleTable/ScheduleTable.tsx
@@ -1,9 +1,6 @@
 import {
   Box,
   Button,
-  Flex,
-  Grid,
-  GridItem,
   Popover,
   PopoverArrow,
   PopoverBody,
@@ -12,12 +9,13 @@ import {
   PopoverTrigger,
   Text,
 } from "@chakra-ui/react";
-import { CellSize, DAY_LABELS, 분 } from "../../constants.ts";
+import { CellSize, DAY_LABELS } from "../../constants.ts";
 import { Schedule } from "../../types.ts";
-import { fill2, parseHnM } from "../../utils.ts";
 import { useDndContext, useDraggable } from "@dnd-kit/core";
 import { CSS } from "@dnd-kit/utilities";
-import { ComponentProps, Fragment, memo, useCallback, useMemo } from "react";
+import { ComponentProps, memo, useCallback, useMemo } from "react";
+import { ScheduleGrid } from "./ScheduleGrid.tsx";
+import { useAutoCallback } from "../../hooks/useAutoCallback.ts";
 
 interface Props {
   tableId: string;
@@ -25,18 +23,6 @@ interface Props {
   onScheduleTimeClick?: (timeInfo: { day: string; time: number }) => void;
   onDeleteButtonClick?: (timeInfo: { day: string; time: number }) => void;
 }
-
-const TIMES = [
-  ...Array(18)
-    .fill(0)
-    .map((v, k) => v + k * 30 * 분)
-    .map((v) => `${parseHnM(v)}~${parseHnM(v + 30 * 분)}`),
-
-  ...Array(6)
-    .fill(18 * 30 * 분)
-    .map((v, k) => v + k * 55 * 분)
-    .map((v) => `${parseHnM(v)}~${parseHnM(v + 50 * 분)}`),
-] as const;
 
 const ScheduleTable = ({
   tableId,
@@ -61,84 +47,19 @@ const ScheduleTable = ({
     [dndContext.active]
   );
 
+  const handleScheduleTimeClick = useAutoCallback(
+    (timeInfo: { day: string; time: number }) => {
+      onScheduleTimeClick?.(timeInfo);
+    }
+  );
+
   return (
     <Box
       position="relative"
       outline={activeTableId === tableId ? "5px dashed" : undefined}
       outlineColor="blue.300"
     >
-      <Grid
-        templateColumns={`120px repeat(${DAY_LABELS.length}, ${CellSize.WIDTH}px)`}
-        templateRows={`40px repeat(${TIMES.length}, ${CellSize.HEIGHT}px)`}
-        bg="white"
-        fontSize="sm"
-        textAlign="center"
-        outline="1px solid"
-        outlineColor="gray.300"
-      >
-        <GridItem
-          key="교시"
-          borderColor="gray.300"
-          bg="gray.100"
-        >
-          <Flex
-            justifyContent="center"
-            alignItems="center"
-            h="full"
-            w="full"
-          >
-            <Text fontWeight="bold">교시</Text>
-          </Flex>
-        </GridItem>
-        {DAY_LABELS.map((day) => (
-          <GridItem
-            key={day}
-            borderLeft="1px"
-            borderColor="gray.300"
-            bg="gray.100"
-          >
-            <Flex
-              justifyContent="center"
-              alignItems="center"
-              h="full"
-            >
-              <Text fontWeight="bold">{day}</Text>
-            </Flex>
-          </GridItem>
-        ))}
-        {TIMES.map((time, timeIndex) => (
-          <Fragment key={`시간-${timeIndex + 1}`}>
-            <GridItem
-              borderTop="1px solid"
-              borderColor="gray.300"
-              bg={timeIndex > 17 ? "gray.200" : "gray.100"}
-            >
-              <Flex
-                justifyContent="center"
-                alignItems="center"
-                h="full"
-              >
-                <Text fontSize="xs">
-                  {fill2(timeIndex + 1)} ({time})
-                </Text>
-              </Flex>
-            </GridItem>
-            {DAY_LABELS.map((day) => (
-              <GridItem
-                key={`${day}-${timeIndex + 2}`}
-                borderWidth="1px 0 0 1px"
-                borderColor="gray.300"
-                bg={timeIndex > 17 ? "gray.100" : "white"}
-                cursor="pointer"
-                _hover={{ bg: "yellow.100" }}
-                onClick={() =>
-                  onScheduleTimeClick?.({ day, time: timeIndex + 1 })
-                }
-              />
-            ))}
-          </Fragment>
-        ))}
-      </Grid>
+      <ScheduleGrid onScheduleTimeClick={handleScheduleTimeClick} />
 
       {schedules.map((schedule, index) => (
         <DraggableSchedule
